Выполнил задание 5.12.
