+++ js/data.js
@@ -1,0 +1,42 @@
+export const PLACEMENT_TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+export const CHECKIN_TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+export const CHECKOUT_TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+export const FEATURES_OF_THE_PLACE = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+export const OFFER_PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+export const PRICE_OF_PLACE_MIN = 1000;
+export const PRICE_OF_PLACE_MAX = 100000;
+export const ROMS_MIN = 1;
+export const ROMS_MAX = 8;
+export const GUEST_MIN = 1;
+export const GUEST_MAX = 5;
+export const DESCRIPTION_OFFER = 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. A amet blanditiis debitis dignissimos, distinctio expedita fugit, magni minima mollitia, nam nisi numquam odit officia quod saepe vero voluptatem? Cumque, voluptatem?';
+export const TITLE_OFFER = 'Lorem ipsum dolor sit amet';
+export const SIMILAR_OFFER_COUNT = 10;
